feat: added ft_lword_search
Searches for a specific byte in a longword array.

refactor: updated ft_quickfind to use ft_lword_search

refactor!: started work on const standardisation in string functions

String functions need a serious rethinking as to how 'const' modifier is used.
Currently there is no concensus as to what method should return an un-modifianle
string or as to how string are treated in libft: is their mutation down to the
user ? Should they be immutable ?