feat: added Dequeue data structure
Resisable array implementation of Dequeue data structure. For n initial elements added to the queue, its array will
be the size of the closes power of 2 n', s.t n' > n. Initial elements are copied to the center of this new array, which
allows to add new elements at the bottom and the top of the queue easily. If there is no space to add new elements in
either direction, will attempt to reorganise Dequeue to make space by centering the elements. Otherwise, doubles the
size of the Dequeue array and copies over all elements to the center of the new array.